Logout button added
Currently Unilogin logout is supported
