+++ app/auth/logout/route.ts
@@ -1,41 +1,29 @@
 import { cookies } from "next/headers"
 import * as client from "openid-client"
 
-import goConfig from "@/lib/config/goConfig"
 import { getUniloginClientConfig } from "@/lib/session/oauth/uniloginClient"
 import { getSession } from "@/lib/session/session"
 
+import {
+  destroySessionAndRedirectToFrontPage,
+  handleUniloginLogout,
+  redirectToFrontPage,
+} from "./helpers"
+
 export async function GET() {
   const session = await getSession()
-  const config = await getUniloginClientConfig()
-  const appUrl = new URL(String(goConfig("app.url")))
 
-  if (!config) {
-    return Response.redirect(appUrl)
+  // If the session is of type anonymous or for some reason is unset,
+  // we don't need to do anything because the user is not logged in.
+  if (!session.type || session.type === "anonymous") {
+    return redirectToFrontPage()
   }
 
-  session.destroy()
-
-  // TODO: Distinguish between session types here.
-  const id_token = (await cookies()).get("go-session:id_token")?.value
-  // TODO: Is this where we want to redirect to if id token cannot be resolved?
-  if (!id_token) {
-    return Response.redirect(appUrl)
-  }
-  const endSessionEndpoint = config.serverMetadata().end_session_endpoint
-
-  if (!endSessionEndpoint) {
-    return Response.redirect(appUrl)
+  if (session.type === "unilogin") {
+    return handleUniloginLogout(session)
   }
 
-  const endSessionUrl = client.buildEndSessionUrl(config, {
-    id_token_hint: id_token,
-  })
-
-  // End session in Unilogin SSO.
-  await fetch(endSessionUrl)
-
-  return Response.redirect(`${appUrl.toString()}?reload-session=true`)
+  return destroySessionAndRedirectToFrontPage(session)
 }
 
 export const dynamic = "force-dynamic"
