+++ app/user/profile/LogoutButton.tsx
@@ -3,9 +3,16 @@
 import { useRouter } from "next/navigation"
 
 import { Button } from "@/components/shared/button/Button"
+import useSession from "@/hooks/useSession"
 
-const UniloginLogoutButton = () => {
+const LogoutButton = () => {
+  const session = useSession()
   const router = useRouter()
+
+  if (!session?.session?.isLoggedIn) {
+    return null
+  }
+
   return (
     <>
       <Button onClick={() => router.push("/auth/logout")} ariaLabel="Log ud">
@@ -14,4 +21,4 @@ const UniloginLogoutButton = () => {
     </>
   )
 }
-export default UniloginLogoutButton
+export default LogoutButton
